refactor(posts): extract generation engine into PostGenerationService (#691)
Move the AI generation engine out of the 2,111-line posts-operations
controller into a dedicated PostGenerationService, mirroring the existing
posts analytics split. The controller drops from 13 to 8 injected
dependencies and becomes a thin HTTP boundary.

- New PostGenerationService owns the account-content and thread-expansion
  background pipelines, post enhancement, hook-variation generation, and the
  shared prompt-building / tweet-parsing helpers.
- Controller endpoints delegate via startAccountContentGeneration,
  expandThreadAsync, enhanceDescription, generateHookVariations, and
  extractLabelFromTweet; HTTP validation, access checks, serialization, and
  request helpers stay on the controller.
- Delete dead generateTweetsAsync / generateThreadAsync (never wired to a
  route; the generations / thread-generations endpoints delegate to
  generateAccountContent -> generateAccountContentAsync).
- Behavior-preserving: routes, DTOs, status codes, error wrapping, async
  fire-and-forget semantics, and websocket/activity side effects unchanged.
- Relocate the three controller-spec tests that reached into moved privates
  into a new post-generation.service.spec.ts; register the real service in the
  controller spec so endpoint assertions still exercise it.
